+++ components/director/internal/domain/runtime/resolver.go
@@ -222,10 +222,12 @@ func (r *Resolver) DeleteRuntime(ctx context.Context, id string) (*graphql.Runti
 	}
 
 	for _, auth := range auths {
-		if auth.Value.Credential.Oauth != nil {
-			err := r.oAuth20Svc.DeleteClientCredentials(ctx, auth.Value.Credential.Oauth.ClientID)
-			if err != nil {
-				return nil, errors.Wrap(err, "while deleting OAuth 2.0 client")
+		if auth.Value != nil {
+			if auth.Value.Credential.Oauth != nil {
+				err := r.oAuth20Svc.DeleteClientCredentials(ctx, auth.Value.Credential.Oauth.ClientID)
+				if err != nil {
+					return nil, errors.Wrap(err, "while deleting OAuth 2.0 client")
+				}
 			}
 		}
 	}
